requirement_shaif.txt
affine==2.4.0
amqp==5.2.0
asgiref==3.7.2
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
bcrypt==4.0.1
billiard==4.2.0
bqplot==0.12.42
branca==0.6.0
Brotli==1.1.0
cachetools==5.3.1
cdo==1.6.0
celery==5.4.0
celery-progress==0.3
certifi==2023.7.22
cffi==1.16.0
cftime==1.6.3
charset-normalizer==3.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
colour==0.1.5
comm==0.1.4
contourpy==1.1.1
cryptography==41.0.4
cssselect2==0.7.0
cycler==0.12.1
decorator==5.1.1
diff-match-patch==20230430
Django==4.1
django-admin-rangefilter==0.13.3
django-celery-results==2.5.1
django-cors-headers==4.2.0
django-filter==23.2
django-import-export==4.0.0
django-pagination==1.0.7
django-pandas==0.6.6
django-redis==5.4.0
django-simple-pagination==1.3
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
earthengine-api==0.1.375
eerepr==0.0.4
et_xmlfile==2.0.0
exceptiongroup==1.1.3
executing==2.0.0
Fiona==1.9.4.post1
folium==0.14.0
fonttools==4.43.1
future==0.18.3
geemap==0.29.1
geocoder==1.38.1
geojson==3.1.0
geojsoncontour==0.4.0
geopandas==0.14.0
google-api-core==2.12.0
google-api-python-client==2.104.0
google-auth==2.23.3
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.2.2
google-cloud-core==2.3.3
google-cloud-storage==2.12.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
greenlet==3.0.0
gunicorn==22.0.0
html5lib==1.1
httplib2==0.22.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipyleaflet==0.17.4
ipython==8.16.1
ipytree==0.2.2
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
kombu==5.3.7
lxml==5.2.2
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mysql-connector-python==8.4.0
netCDF4==1.6.5
numpy==1.25.2
oauthlib==3.3.1
openpyxl==3.1.5
packaging==23.2
panda==0.3.1
pandas==2.1.0
paramiko==3.3.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
plotly==5.17.0
prompt-toolkit==3.0.39
proto-plus==1.26.1
protobuf==4.24.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydyf==0.10.0
Pygments==2.16.1
# Editable install with no version control (pyidw==0.2.20)
-e /home/rimes/ffwc-rebase/backend/ffwc_django_project/pyidw-0.2.20
PyJWT==1.7.1
PyMySQL==1.1.0
PyNaCl==1.5.0
pyogrio==0.11.0
pyparsing==3.1.1
pyperclip==1.8.2
pyphen==0.15.0
pyproj==3.6.1
pyscissor==1.1.7
pyshp==2.3.1
python-box==7.1.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
rasterio==1.3.8
ratelim==0.1.6
redis==5.0.4
requests==2.31.0
requests-oauthlib==2.0.0
rioxarray==0.15.0
rsa==4.9
scikit-learn==1.4.2
scipy==1.11.3
scooby==0.9.2
shapely==2.0.1
six==1.16.0
sklearn==0.0
snuggs==1.4.7
SQLAlchemy==2.0.20
sqlparse==0.4.4
sshtunnel==0.4.0
stack-data==0.6.3
tablib==3.5.0
tenacity==8.2.3
termcolor==2.3.0
threadpoolctl==3.2.0
tinycss2==1.3.0
tinyhtml5==2.0.0
tqdm==4.66.4
traitlets==5.11.2
traittypes==0.2.1
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
vine==5.1.0
wcwidth==0.2.8
weasyprint==62.1
webencodings==0.5.1
whitenoise==6.5.0
widgetsnbextension==4.0.9
xarray==2024.3.0
xyzservices==2023.10.0
yaspin==3.0.2
zipp==3.17.0
zopfli==0.2.3
